requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-google-genai
google-generativeai
matplotlib
pandas
numpy
streamlit
plotly
jupyter